requirements.txt
pytest>=7.2.0
selene>=2.0.0b16
python-dotenv==0.20.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
